Add files to generate and display animation
Add HTML file for actually displaying the animation.

Add CSS file for more modularized styling

Add JS file for all of the heavy lifting.
This does all of the animation work.

TODO: Currently the color of one of the blocks
changes during any time that the other block is in
contact with it.  This needs to be modified such that
it will only change once per contact.